ml/requirements2.txt
fastapi
uvicorn[standard]
APScheduler
scikit-learn
SQLAlchemy
psycopg2-binary
python-dotenv
